3.1-drf-intro/smart_home/requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow